Experiments with unusual splitscreen distances
Work done in zac tab; makes it so the splitscreen distance changes dynamically based on the direction the players are moving away from eachother, thus making the transition smoother. Default distance for the NewCamMan script is default set to 4.5 in the vertical and 8 in the horizontal, which is optimal for a camera with a size of 4.5 and an aspect ration of 16:9.